pyproject.toml
[project]
name = "ncc-dashboard-bc"
version = "0.1.0"
description = ""
authors = [
    {name = "officialalkenes",email = "bello.abdul@ivendigital.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
